🌱 Refactor: split CardWrapper.tsx into focused sub-components
Extract 4 sub-components from the 1682-line CardWrapper monolith:

- CardActionMenu.tsx: three-dot menu portal + resize/height submenus
  (~365 lines, 10 state variables, 5 useEffects isolated)
- InfoTooltip.tsx: portal tooltip + shared Escape-key coordinator
  (~155 lines, module-level escape stack scoped to file)
- PendingSwapNotification.tsx: swap countdown banner
  (~78 lines, pure presentational)
- InstallCTAFlow.tsx: demo install CTA + cluster selection + prompt
  confirm + manual guide modal (~150 lines, 3 state variables)

CardWrapper.tsx: 1682 → 1004 lines (−40%).

Menu state (showMenu, showResizeMenu, showHeightMenu, menuPosition,
resizeMenuOnLeft, heightMenuOnLeft, etc.) is now isolated in
CardActionMenu, preventing unnecessary re-renders of the card body
when the menu opens/closes.

No behavioral changes — all props, aria attributes, data-tour
attributes, and portal rendering preserved exactly.
